✨ Add notification settings to user settings and refactor notification logic for task owners
